+++ CHANGELOG.md
@@ -11,6 +11,16 @@ Changelog
 
 _Note: Gaps between patch versions are faulty, broken or test releases._
 
+## v3.7.0 (2020-03-17)
+
+#### :rocket: New Feature
+
+* Added `parent` to `core/object/watch/interface/WatchHandlerParams`
+
+#### :bug: Bug Fix
+
+* Fixed `core/object/watch` collapsing of events
+
 ## v3.6.1 (2020-03-16)
 
 #### :bug: Bug Fix